Very basic splitting of code into seperate files.
